Rename Database.java files to distinguish the todos and user databases
The idea for this edit is based on a comment by Paul on Slack. 
"Yes, I believe that is the intent. In fact you may want to name them more specifically, eg TodosDatabase"
