+++ server/src/main/java/umm3601/Server.java
@@ -4,24 +4,24 @@ import java.io.IOException;
 
 import io.javalin.Javalin;
 import io.javalin.http.staticfiles.Location;
-//import umm3601.user.Database;
+import umm3601.user.UserDatabase;
 import umm3601.user.UserController;
 import umm3601.todo.TodoController;
-import umm3601.todo.Database;
+import umm3601.todo.TodoDatabase;
 
 public class Server {
 
   public static final String CLIENT_DIRECTORY = "../client";
   public static final String USER_DATA_FILE = "/users.json";
   public static final String TODOS_DATA_FILE = "/todos.json";
-  private static Database userDatabase;
-  private static Database todoDatabase;
+  private static UserDatabase userDatabase;
+  private static TodoDatabase todoDatabase;
 
 
   public static void main(String[] args) {
 
     // Initialize dependencies
-   // UserController userController = buildUserController();
+    UserController userController = buildUserController();
     TodoController todoController = buildTodoController();
 
     Javalin server = Javalin.create(config -> {
@@ -41,13 +41,14 @@ public class Server {
     // API endpoints
 
     // Get specific user
-    // server.get("api/users/:id", ctx -> userController.getUser(ctx));
+    server.get("api/users/:id", ctx -> userController.getUser(ctx));
 
-    // // List users, filtered using query parameters
-    // server.get("api/users", ctx -> userController.getUsers(ctx));
+     // List users, filtered using query parameters
+     server.get("api/users", ctx -> userController.getUsers(ctx));
 
     server.get("api/todos/:id", ctx -> todoController.getTodo(ctx));
     server.get("api/todos", ctx -> todoController.getTodos(ctx));
+    //server.get("api/todos?limit=", ctx -> todoController.getTodos);
   }
 
   /***
@@ -58,28 +59,28 @@ public class Server {
    * reading from the JSON "database" file. If that happens we'll print out an
    * error message exit the program.
    */
-  // private static UserController buildUserController() {
-  //   UserController userController = null;
+   private static UserController buildUserController() {
+   UserController userController = null;
 
-  //   try {
-  //     userDatabase = new Database(USER_DATA_FILE);
-  //     userController = new UserController(userDatabase);
-  //   } catch (IOException e) {
-  //     System.err.println("The server failed to load the user data; shutting down.");
-  //     e.printStackTrace(System.err);
+     try {
+      userDatabase = new UserDatabase(USER_DATA_FILE);
+       userController = new UserController(userDatabase);
+     } catch (IOException e) {
+       System.err.println("The server failed to load the user data; shutting down.");
+       e.printStackTrace(System.err);
 
-  //     // Exit from the Java program
-  //     System.exit(1);
-  //   }
+       // Exit from the Java program
+       System.exit(1);
+     }
 
-  //   return userController;
-  // }
+     return userController;
+   }
 
   private static TodoController buildTodoController() {
     TodoController todoController = null;
 
     try {
-      todoDatabase = new Database(TODOS_DATA_FILE);
+      todoDatabase = new TodoDatabase(TODOS_DATA_FILE);
       todoController = new TodoController(todoDatabase);
     } catch (IOException e) {
       System.err.println("The server failed to load the user data; shutting down.");
