+++ server/src/main/java/umm3601/user/UserController.java
@@ -8,7 +8,7 @@ import io.javalin.http.NotFoundResponse;
  */
 public class UserController {
 
-  private Database database;
+  private UserDatabase database;
 
   /**
    * Construct a controller for users.
@@ -19,7 +19,7 @@ public class UserController {
    *
    * @param database the `Database` containing user data
    */
-  public UserController(Database database) {
+  public UserController(UserDatabase database) {
     this.database = database;
   }
