+++ server/src/test/java/umm3601/user/FilterUsersByAgeFromDB.java
@@ -18,7 +18,7 @@ public class FilterUsersByAgeFromDB {
 
   @Test
   public void filterUsersByAge() throws IOException {
-    Database db = new Database("/users.json");
+    UserDatabase db = new UserDatabase("/users.json");
     User[] allUsers = db.listUsers(new HashMap<>());
 
     User[] age27Users = db.filterUsersByAge(allUsers, 27);
@@ -30,7 +30,7 @@ public class FilterUsersByAgeFromDB {
 
   @Test
   public void listUsersWithAgeFilter() throws IOException {
-    Database db = new Database("/users.json");
+    UserDatabase db = new UserDatabase("/users.json");
     Map<String, List<String>> queryParams = new HashMap<>();
 
     queryParams.put("age", Arrays.asList(new String[] { "27" }));
